Refine gpiochip discovery
There are moves afoot [1] to change the gpiochip device on the Pi 5 from
gpiochip4 to gpiochip0. The current gpiochip discovery in rpi-lgpio is
crude, to say the least. This commit refines the method to discovering
the chip based on the GPIO driver in use (which should work on all Pi
boards with all kernels).

Closes: #10

[1]: https://github.com/raspberrypi/linux/pull/6144