Force upgrades for old lgpio
This may or may not work, but I'm hoping it'll avoid more people
encountering errors with old versions of lgpio. The version shipped in
Ubuntu jammy is actually 0.1.6.1 but has an internal version string of
0.1.0.1. Meanwhile, the old (incompatible) version on PyPI is 0.1.0.0,
hence I *hope* this will be sufficient to force upgrades where necessary
but retain compatibility with old releases in OS'.